+++ Engine/Renderer/GLES31/GfxTextureCubeMap.cpp
@@ -212,7 +212,7 @@ bool CGfxTextureCubeMap::TransferTexture2D(int face, const gli::texture2d *textu
 	return true;
 }
 
-bool CGfxTextureCubeMap::TransferTexture2D(int face, int level, uint32_t format, int xoffset, int yoffset, int width, int height, uint32_t type, int size, const void *data)
+bool CGfxTextureCubeMap::TransferTexture2D(int face, int level, uint32_t format, int xoffset, int yoffset, int width, int height, uint32_t type, uint32_t size, const void *data)
 {
 	if (m_texture == 0) {
 		return false;
@@ -230,7 +230,7 @@ bool CGfxTextureCubeMap::TransferTexture2D(int face, int level, uint32_t format,
 	{
 		CGfxProfiler::DecTextureDataSize(m_size[face][level]);
 		{
-			m_size[face][level] = (uint32_t)size;
+			m_size[face][level] = size;
 			glTexSubImage2D(face, level, xoffset, yoffset, width, height, format, type, data);
 		}
 		CGfxProfiler::IncTextureDataSize(m_size[face][level]);
@@ -240,7 +240,7 @@ bool CGfxTextureCubeMap::TransferTexture2D(int face, int level, uint32_t format,
 	return true;
 }
 
-bool CGfxTextureCubeMap::TransferTexture2DCompressed(int face, int level, uint32_t format, int xoffset, int yoffset, int width, int height, int size, const void *data)
+bool CGfxTextureCubeMap::TransferTexture2DCompressed(int face, int level, uint32_t format, int xoffset, int yoffset, int width, int height, uint32_t size, const void *data)
 {
 	if (m_texture == 0) {
 		return false;
@@ -258,7 +258,7 @@ bool CGfxTextureCubeMap::TransferTexture2DCompressed(int face, int level, uint32
 	{
 		CGfxProfiler::DecTextureDataSize(m_size[face][level]);
 		{
-			m_size[face][level] = (uint32_t)size;
+			m_size[face][level] = size;
 			glCompressedTexSubImage2D(face, level, xoffset, yoffset, width, height, format, size, data);
 		}
 		CGfxProfiler::IncTextureDataSize(m_size[face][level]);
